The encode test works in test file but fails in the tool

+++ src/caesarChipher.h
@@ -2,7 +2,7 @@
 #include "pxx.h"
 #include <Tools/Tools.h>
 
-void caesarEncoding(const int keyValue, const std::string &inputData);
-void ceasarDecoding(const int keyValue, const std::string &inputData);
+void caesarEncoding(const uint_fast8_t keyValue, const std::string &inputData);
+void ceasarDecoding(const uint_fast8_t keyValue, const std::string &inputData);
 
 void ceasarDecodingBruteforce(const std::string &data);
